requirements.txt
pandas~=2.2.1
numpy~=1.26.2
scikit-learn~=1.4.1.post1
Pillow~=10.1.0